Split up tedious
