+++ lib/tedious/index.js
@@ -1,0 +1,28 @@
+'use struct'
+
+const base = require('../base')
+const ConnectionPool = require('./connection-pool')
+const Transaction = require('./transaction')
+const Request = require('./request')
+
+module.exports = Object.assign({
+  ConnectionPool,
+  Transaction,
+  Request,
+  PreparedStatement: base.PreparedStatement
+}, base.exports)
+
+Object.defineProperty(module.exports, 'Promise', {
+  enumerable: true,
+  get: () => {
+    return base.Promise
+  },
+  set: (value) => {
+    base.Promise = value
+  }
+})
+
+base.driver.name = 'tedious'
+base.driver.ConnectionPool = ConnectionPool
+base.driver.Transaction = Transaction
+base.driver.Request = Request
