+++ lib/tedious/request.js
@@ -2,19 +2,16 @@
 
 const tds = require('tedious')
 const debug = require('debug')('mssql:tedi')
-
-const base = require('./base')
-const TYPES = require('./datatypes').TYPES
-const declare = require('./datatypes').declare
-const cast = require('./datatypes').cast
-const DECLARATIONS = require('./datatypes').DECLARATIONS
-const UDT = require('./udt').PARSERS
-const Table = require('./table')
-const IDS = require('./utils').IDS
-const objectHasProperty = require('./utils').objectHasProperty
+const BaseRequest = require('../base/request')
+const RequestError = require('../error/request-error')
+const { IDS, objectHasProperty } = require('../utils')
+const { TYPES, DECLARATIONS, declare, cast } = require('../datatypes')
+const Table = require('../table')
+const { PARSERS: UDT } = require('../udt')
 
 const JSON_COLUMN_ID = 'JSON_F52E2B61-18A1-11d1-B105-00805F49916B'
 const XML_COLUMN_ID = 'XML_F52E2B61-18A1-11d1-B105-00805F49916B'
+
 const N_TYPES = {
   BitN: 0x68,
   DateTimeN: 0x6F,
@@ -191,212 +188,7 @@ const parameterCorrection = function (value) {
   }
 }
 
-class ConnectionPool extends base.ConnectionPool {
-  _poolCreate () {
-    return new base.Promise((resolve, reject) => {
-      const resolveOnce = (v) => {
-        resolve(v)
-        resolve = reject = () => {}
-      }
-      const rejectOnce = (e) => {
-        reject(e)
-        resolve = reject = () => {}
-      }
-      const cfg = {
-        server: this.config.server,
-        options: Object.assign({
-          encrypt: typeof this.config.encrypt === 'boolean' ? this.config.encrypt : false
-        }, this.config.options),
-        authentication: Object.assign({
-          type: this.config.domain !== undefined ? 'ntlm' : 'default',
-          options: {
-            userName: this.config.user,
-            password: this.config.password,
-            domain: this.config.domain
-          }
-        }, this.config.authentication)
-      }
-
-      cfg.options.database = this.config.database
-      cfg.options.port = this.config.port
-      cfg.options.connectTimeout = this.config.connectionTimeout || this.config.timeout || 15000
-      cfg.options.requestTimeout = this.config.requestTimeout != null ? this.config.requestTimeout : 15000
-      cfg.options.tdsVersion = cfg.options.tdsVersion || '7_4'
-      cfg.options.rowCollectionOnDone = false
-      cfg.options.rowCollectionOnRequestCompletion = false
-      cfg.options.useColumnNames = false
-      cfg.options.appName = cfg.options.appName || 'node-mssql'
-
-      // tedious always connect via tcp when port is specified
-      if (cfg.options.instanceName) delete cfg.options.port
-
-      if (isNaN(cfg.options.requestTimeout)) cfg.options.requestTimeout = 15000
-      if (cfg.options.requestTimeout === Infinity) cfg.options.requestTimeout = 0
-      if (cfg.options.requestTimeout < 0) cfg.options.requestTimeout = 0
-
-      if (this.config.debug) {
-        cfg.options.debug = {
-          packet: true,
-          token: true,
-          data: true,
-          payload: true
-        }
-      }
-      const tedious = new tds.Connection(cfg)
-      IDS.add(tedious, 'Connection')
-      debug('pool(%d): connection #%d created', IDS.get(this), IDS.get(tedious))
-      debug('connection(%d): establishing', IDS.get(tedious))
-
-      tedious.once('connect', err => {
-        if (err) {
-          err = new base.ConnectionError(err)
-          return rejectOnce(err)
-        }
-
-        debug('connection(%d): established', IDS.get(tedious))
-        resolveOnce(tedious)
-      })
-
-      tedious.on('end', () => {
-        const err = new base.ConnectionError('The connection ended without ever completing the connection')
-        rejectOnce(err)
-      })
-      tedious.on('error', err => {
-        if (err.code === 'ESOCKET') {
-          tedious.hasError = true
-        } else {
-          this.emit('error', err)
-        }
-        rejectOnce(err)
-      })
-
-      if (this.config.debug) {
-        tedious.on('debug', this.emit.bind(this, 'debug', tedious))
-      }
-      if (typeof this.config.beforeConnect === 'function') {
-        this.config.beforeConnect(tedious)
-      }
-    })
-  }
-
-  _poolValidate (tedious) {
-    return tedious && !tedious.closed && !tedious.hasError
-  }
-
-  _poolDestroy (tedious) {
-    return new base.Promise((resolve, reject) => {
-      if (!tedious) {
-        resolve()
-        return
-      }
-      debug('connection(%d): destroying', IDS.get(tedious))
-
-      if (tedious.closed) {
-        debug('connection(%d): already closed', IDS.get(tedious))
-        resolve()
-      } else {
-        tedious.once('end', () => {
-          debug('connection(%d): destroyed', IDS.get(tedious))
-          resolve()
-        })
-
-        tedious.close()
-      }
-    })
-  }
-}
-
-class Transaction extends base.Transaction {
-  constructor (parent) {
-    super(parent)
-
-    this._abort = () => {
-      if (!this._rollbackRequested) {
-        // transaction interrupted because of XACT_ABORT
-
-        const pc = this._acquiredConnection
-
-        // defer releasing so connection can switch from SentClientRequest to LoggedIn state
-        setImmediate(this.parent.release.bind(this.parent), pc)
-
-        this._acquiredConnection.removeListener('rollbackTransaction', this._abort)
-        this._acquiredConnection = null
-        this._acquiredConfig = null
-        this._aborted = true
-
-        this.emit('rollback', true)
-      }
-    }
-  }
-
-  _begin (isolationLevel, callback) {
-    super._begin(isolationLevel, err => {
-      if (err) return callback(err)
-
-      debug('transaction(%d): begin', IDS.get(this))
-
-      this.parent.acquire(this, (err, connection, config) => {
-        if (err) return callback(err)
-
-        this._acquiredConnection = connection
-        this._acquiredConnection.on('rollbackTransaction', this._abort)
-        this._acquiredConfig = config
-
-        connection.beginTransaction(err => {
-          if (err) err = new base.TransactionError(err)
-
-          debug('transaction(%d): begun', IDS.get(this))
-
-          callback(err)
-        }, this.name, this.isolationLevel)
-      })
-    })
-  }
-
-  _commit (callback) {
-    super._commit(err => {
-      if (err) return callback(err)
-
-      debug('transaction(%d): commit', IDS.get(this))
-
-      this._acquiredConnection.commitTransaction(err => {
-        if (err) err = new base.TransactionError(err)
-
-        this._acquiredConnection.removeListener('rollbackTransaction', this._abort)
-        this.parent.release(this._acquiredConnection)
-        this._acquiredConnection = null
-        this._acquiredConfig = null
-
-        if (!err) debug('transaction(%d): commited', IDS.get(this))
-
-        callback(err)
-      })
-    })
-  }
-
-  _rollback (callback) {
-    super._rollback(err => {
-      if (err) return callback(err)
-
-      debug('transaction(%d): rollback', IDS.get(this))
-
-      this._acquiredConnection.rollbackTransaction(err => {
-        if (err) err = new base.TransactionError(err)
-
-        this._acquiredConnection.removeListener('rollbackTransaction', this._abort)
-        this.parent.release(this._acquiredConnection)
-        this._acquiredConnection = null
-        this._acquiredConfig = null
-
-        if (!err) debug('transaction(%d): rolled back', IDS.get(this))
-
-        callback(err)
-      })
-    })
-  }
-}
-
-class Request extends base.Request {
+class Request extends BaseRequest {
   /*
   Execute specified sql batch.
   */
@@ -417,11 +209,11 @@ class Request extends base.Request {
       table._makeBulk()
 
       if (!table.name) {
-        return callback(new base.RequestError('Table name must be specified for bulk insert.', 'ENAME'))
+        return callback(new RequestError('Table name must be specified for bulk insert.', 'ENAME'))
       }
 
       if (table.name.charAt(0) === '@') {
-        return callback(new base.RequestError("You can't use table variables for bulk insert.", 'ENAME'))
+        return callback(new RequestError("You can't use table variables for bulk insert.", 'ENAME'))
       }
 
       const errors = []
@@ -431,7 +223,7 @@ class Request extends base.Request {
       const handleError = (doReturn, connection, info) => {
         let err = new Error(info.message)
         err.info = info
-        err = new base.RequestError(err, 'EREQUEST')
+        err = new RequestError(err, 'EREQUEST')
 
         if (this.stream) {
           this.emit('error', err)
@@ -474,7 +266,7 @@ class Request extends base.Request {
         if (this.canceled) {
           debug('request(%d): canceled', IDS.get(this))
           this.parent.release(connection)
-          return callback(new base.RequestError('Canceled.', 'ECANCEL'))
+          return callback(new RequestError('Canceled.', 'ECANCEL'))
         }
 
         this._cancel = () => {
@@ -490,7 +282,7 @@ class Request extends base.Request {
         const done = (err, rowCount) => {
           // to make sure we handle no-sql errors as well
           if (err && (!errors.length || (errors.length && err.message !== errors[errors.length - 1].message))) {
-            err = new base.RequestError(err, 'EREQUEST')
+            err = new RequestError(err, 'EREQUEST')
             if (this.stream) this.emit('error', err)
             errors.push(err)
           }
@@ -571,7 +363,7 @@ class Request extends base.Request {
       const handleError = (doReturn, connection, info) => {
         let err = new Error(info.message)
         err.info = info
-        err = new base.RequestError(err, 'EREQUEST')
+        err = new RequestError(err, 'EREQUEST')
 
         if (this.stream) {
           this.emit('error', err)
@@ -616,7 +408,7 @@ class Request extends base.Request {
         if (this.canceled) {
           debug('request(%d): canceled', IDS.get(this))
           this.parent.release(connection)
-          return callback(new base.RequestError('Canceled.', 'ECANCEL'))
+          return callback(new RequestError('Canceled.', 'ECANCEL'))
         }
 
         this._cancel = () => {
@@ -634,7 +426,7 @@ class Request extends base.Request {
         const req = new tds.Request(command, err => {
           // to make sure we handle no-sql errors as well
           if (err && (!errors.length || (errors.length && err.message !== errors[errors.length - 1].message))) {
-            err = new base.RequestError(err, 'EREQUEST')
+            err = new RequestError(err, 'EREQUEST')
             if (this.stream) this.emit('error', err)
             errors.push(err)
           }
@@ -702,8 +494,7 @@ class Request extends base.Request {
               this.emit('recordset', columns)
             }
           }
-        }
-        )
+        })
 
         const doneHandler = (rowCount, more) => {
           if (rowCount != null) rowsAffected.push(rowCount)
@@ -721,7 +512,7 @@ class Request extends base.Request {
                 }
               } catch (ex) {
                 row = null
-                const ex2 = new base.RequestError(new Error(`Failed to parse incoming JSON. ${ex.message}`), 'EJSON')
+                const ex2 = new RequestError(new Error(`Failed to parse incoming JSON. ${ex.message}`), 'EJSON')
 
                 if (this.stream) this.emit('error', ex2)
 
@@ -819,7 +610,7 @@ class Request extends base.Request {
               let value = getTediousType(param.type).validate(param.value)
 
               if (value instanceof TypeError) {
-                value = new base.RequestError(`Validation failed for parameter '${name}'. ${value.message}`, 'EPARAM')
+                value = new RequestError(`Validation failed for parameter '${name}'. ${value.message}`, 'EPARAM')
 
                 this.parent.release(connection)
                 return callback(value)
@@ -908,7 +699,7 @@ class Request extends base.Request {
       const handleError = (doReturn, connection, info) => {
         let err = new Error(info.message)
         err.info = info
-        err = new base.RequestError(err, 'EREQUEST')
+        err = new RequestError(err, 'EREQUEST')
 
         if (this.stream) {
           this.emit('error', err)
@@ -953,7 +744,7 @@ class Request extends base.Request {
         if (this.canceled) {
           debug('request(%d): canceled', IDS.get(this))
           this.parent.release(connection)
-          return callback(new base.RequestError('Canceled.', 'ECANCEL'))
+          return callback(new RequestError('Canceled.', 'ECANCEL'))
         }
 
         this._cancel = () => {
@@ -971,7 +762,7 @@ class Request extends base.Request {
         const req = new tds.Request(procedure, err => {
           // to make sure we handle no-sql errors as well
           if (err && (!errors.length || (errors.length && err.message !== errors[errors.length - 1].message))) {
-            err = new base.RequestError(err, 'EREQUEST')
+            err = new RequestError(err, 'EREQUEST')
             if (this.stream) this.emit('error', err)
             errors.push(err)
           }
@@ -1066,7 +857,7 @@ class Request extends base.Request {
                 }
               } catch (ex) {
                 row = null
-                const ex2 = new base.RequestError(new Error(`Failed to parse incoming JSON. ${ex.message}`), 'EJSON')
+                const ex2 = new RequestError(new Error(`Failed to parse incoming JSON. ${ex.message}`), 'EJSON')
 
                 if (this.stream) this.emit('error', ex2)
 
@@ -1148,24 +939,4 @@ class Request extends base.Request {
   }
 }
 
-module.exports = Object.assign({
-  ConnectionPool,
-  Transaction,
-  Request,
-  PreparedStatement: base.PreparedStatement
-}, base.exports)
-
-Object.defineProperty(module.exports, 'Promise', {
-  enumerable: true,
-  get: () => {
-    return base.Promise
-  },
-  set: (value) => {
-    base.Promise = value
-  }
-})
-
-base.driver.name = 'tedious'
-base.driver.ConnectionPool = ConnectionPool
-base.driver.Transaction = Transaction
-base.driver.Request = Request
+module.exports = Request
